requirements.txt
affine==2.3.1
attrs==21.4.0
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorlog==6.6.0
Cython==0.29.30
idna==3.3
image-similarity-measures==0.3.5
imageio==2.19.3
importlib-metadata==4.11.4
mutagen==1.45.1
networkx==2.6.3
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
phasepack==1.5
Pillow==9.1.1
pycryptodomex==3.14.1
pyFFTW==0.13.0
pyparsing==3.0.9
PyWavelets==1.3.0
rasterio==1.2.10
requests==2.28.0
scikit-image==0.19.3
scipy==1.7.3
snuggs==1.4.7
tifffile==2021.11.2
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
vidgear==0.2.5
websockets==10.3
yt-dlp==2022.5.18
zipp==3.8.0
